zfs_vnop_windows cleanup
The mounting code is large enough to get its own file:
zfs_vnop_windows_mount.c

Clean up code standard style

Move many small repetitive helper functions from zfs_vnop_windows.c into
zfs_vnop_windows_lib.c - standardise on style.

+++ ZFSin/spl/include/sys/types.h
@@ -75,7 +75,7 @@ typedef unsigned int pid_t;
 typedef uintptr_t pc_t;
 typedef uint64_t ssize_t;
 typedef uint64_t vm_offset_t;
-typedef uint32_t dev_t;
+typedef uint64_t dev_t;
 #define NGROUPS 16
 
 typedef unsigned short umode_t;
